requirements.txt
docker 
matplotlib
seaborn
pandas
google-adk